updown-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.3
uWSGI==2.0.21
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv
flask-migrate
flask-bcrypt
flask-jwt-extended
boto3
flask-cors